DHP-93 Moved the section components into their own files; Add test id to section components.
